3rd Commit
This commit contains the code documentation and instructions to install. The bot can now display the followers and the accounts being followed by the bot account. 40 new tweet's containing the tag "#IoT" will be "liked" by the bot.
The bot can now monitor @boltiot's twitter feed and will like any tweets that are posted when it is running.